% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_2_variance_curves-ElectrJStat15.R
\name{comp2condvar}
\alias{comp2condvar}
\title{comp2condvar}
\usage{
comp2condvar(x1, y1, x2, y2, B = 1000, bandwidths = "cv", sigma.w = 1)
}
\arguments{
\item{x1}{TODO}

\item{y1}{TODO}

\item{x2}{TODO}

\item{y2}{TODO}

\item{B}{bootstrap samples}

\item{bandwidths}{TODO}

\item{sigma.w}{TODO}
}
\description{
TODO
}
