% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/comparison_2_regression_curves_onesided-JSPI09.R
\name{comp2regr.onesided}
\alias{comp2regr.onesided}
\title{comp2regr.onesided}
\usage{
comp2regr.onesided(x1, x2, y1, y2, p = function(x) {     0.5 },
  w1 = function(x) {     1 }, w2 = function(x) {     1 })
}
\description{
function that calculates the p-value based on c-v bandwidths
}
